+++ app/src/data/DataMenu.tsx
@@ -1,5 +1,5 @@
 import { User } from "@isa2026/api/src/utils/dbtypes.ts";
-import { LinkButton } from "../components/LinkButton/LinkButton.tsx";
+import { LinkButton } from "../components/Button/LinkButton/LinkButton.tsx";
 import styles from "./DataMenu.module.css";
 
 type DataMenuProps = {
