+++ app/src/scout/ScoutLayout.tsx
@@ -11,7 +11,7 @@ import { Box, Button, Stack, Tab, Tabs } from "@mui/material";
 import EventEmitter from "events";
 import { useEffect, useRef, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { ScoutPageContainer } from "../components/ScoutPageContainer/ScoutPageContainer.tsx";
+import { ScoutPageContainer } from "../components/PageContainer/ScoutPageContainer/ScoutPageContainer.tsx";
 import { DeviceSetupObj } from "../setup/DeviceSetup.tsx";
 import { getDBHumanPlayerEntries, putDBEntry } from "../utils/idb.ts";
 import { trpc } from "../utils/trpc.ts";
