+++ app/src/setup/CreateEvent.tsx
@@ -24,7 +24,7 @@ import {
 import { Overwrite } from "@trpc/server/unstable-core-do-not-import";
 import { useState } from "react";
 import { Fragment } from "react/jsx-runtime";
-import { Th } from "../components/Table.tsx";
+import { Th } from "../components/Table/Table.tsx";
 import { putDBEvent, putDBMatches } from "../utils/idb.ts";
 
 //TODO: either improve or delete this
